Fixed approve message for BasicTransfer Transactions
Previously, the if statement in generateTransactionDescription() -> switch -> case TransactionMetadataBasicTransfer was always evaluating to true because === compares reference and not equality. They are different objects, so if statement always evaluated to true. It is now changed to check for equality of publicKeys. 

For more info on why this was a problem: https://github.com/bitclout/identity/issues/35