diamondhands cleanup
Biggest thing I did is make it so that the transactionSpendingLimit
bytes are passed directly rather than being computed from the JSON that
comes via the URL params. If we don't do this, we run the risk of
serializing differently in identity vs on backend, which would cause a
very hard-to-debug problem whereby users would have their "approve
derived key" txns rejected for a "bad access hash." The way we have it
now, the spending limit hex that we submit to AuthorizeDerivedKey on
the backend is guaranteed to be identitcal to the hex that we use to
generate the accessHash in identity, which wasn't the case before.