fix: keep track of previous user info for multiple metamask logins (#216)
When a user logs in with metamask we generate a new derived key and
new deposit addresses. Previously we were keeping track of only the
most recent login data and overwriting any previous data with it.
This makes it such that we keep an "archive" of the previous derived
information in case we need it for support in the future.

This also makes it such that we re-use any previously generated
messaging key randomness since this is a deterministic value. This
prevents metamask users from needing to approve a new messaging key
each time they log in.